requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
ascii_colors==0.11.7
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
configparser==7.2.0
distro==1.9.0
dotenv==0.9.9
fastapi==0.128.0
frozenlist==1.8.0
google-api-core==2.29.0
google-auth==2.47.0
google-genai==1.60.0
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
json_repair==0.55.0
lightrag-hku==1.4.9.11
multidict==6.7.0
nano-vectordb==0.0.4.3
networkx==3.6.1
numpy==2.4.1
ollama==0.6.1
packaging==26.0
pandas==2.3.3
pipmaster==1.1.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
rdflib==7.5.0
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
setuptools==80.10.1
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tenacity==9.1.2
tiktoken==0.12.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.3.1
websockets==15.0.1
xlsxwriter==3.2.9
yarl==1.22.0
